Converted double timestamps to 64-bit integers, fixed bug in domain server (was faiing to update timestamps).

+++ audio-mixer/src/main.cpp
@@ -318,7 +318,7 @@ int main(int argc, const char* argv[]) {
             }
         }
         
-        double usecToSleep = usecTimestamp(&startTime) + (++nextFrame * BUFFER_SEND_INTERVAL_USECS) - usecTimestampNow();
+        long long usecToSleep = usecTimestamp(&startTime) + (++nextFrame * BUFFER_SEND_INTERVAL_USECS) - usecTimestampNow();
         
         if (usecToSleep > 0) {
             usleep(usecToSleep);
